Implement stepi enough for (some) breakpoints.
Trying to continue from one breakpoint to another fails, of course :).
This commit adds a breakpoint test which passes when run by hand, but
fails when run in the harness.  The gdb tests have had similar issues
in the past and are now failing, so disable the whole lot until the
harness is proven out.