Track sighandlers established by sigaction.
If we SINGLESTEP to see if we entered a sighandler from some execution
states, like from a blocked SIG_IGN syscall on a non-main thread, or a
core-dumping fatal signal on a non-main thread, we can end up
deadlocking.  What we want to know is whether we /will/ enter a
sighandler, before we step (or whatever) and record the sigframe.
This patch adds that information but only checks it against what we
see from the SINGLESTEP.  A later patch will use the sighandler table
to selectively SINGLESTEP into sighandlers.