Fix syscallbuf initialization.
This includes two separate changes.  First, we switch to using
positive invalid syscall numbers for the magic "rrcalls".  rr doesn't
see traps for the exit of negative syscall numbers on (at least) linux
3.8.0 / eglibc 2.17, so it wasn't able to process the old negative
init-syscallbuf rrcall.

And second, we stop assuming that the members of the msghdr that rr
touches during syscallbuf init are word-aligned.