Fix two bugs in the handling of timeslice interrupts.
The first problem was that, in theory, it's possible for a signal to
arrive before the initial rr tracee process has exec()'d the tracee
executable image.  But as this patch documents in code comments, if
that happens replay can't find the right execution point to deliver
the signal, in general.

The second problem is that due to a logic bug, we would previously
program the hpc interrupt for less than SKID_SIZE rcbs.  This causes,
at least, spurious SIG_TIMESLICE's to arrive, which confuses replay.

There's also some minor cleanup of documentation of the
async_signal_syscalls test, which takes an argument that's the log
base 2 of the number of iterations to make.