Add -Wstrict-prototypes
This patch adds -Wstrict-prototypes to the default compiler flags.

While browing the source I noticed a number of declarations of the
form:

const struct flags* rr_flags();

In C this doesn't do what you'd expect; you have to write "(void)" to
declare a function taking no arguments.

Enabling -Wstrict-prototypes found the instances of this, and it also
caught one bug: the "flags" parameter to "replay" was never used, and
the callers didn't pass it in.