Auto-launch gdb during replay, but only open socket for "replay -s".
This changes the rr process tree.  We reserve the top-level rr process
as a future debugger client.  It forks a child which drives replay
until debugger-connection time.  Then the child notifies the parent,
which exec()s gdb on top of itself (preparing the magic commands to
connect to the remote target).

The previous behavior can be re-enabled by passing "-s".  This is
useful when you're frequently restarting execution with nontrivial
breakpoints, and you want them preserved across restarts by detaching
and re-attaching to a new run.