Simply the default record/replay workflow.
With this change, traces are saved to ~/.rr/ by default (this can
still be overridden with the _RR_TRACE_DIR env var.)  In addition, a
"latest-trace" symlink is maintained within the saved-trace dir, and
updated with each invocation of |rr record|.  This enables the next
simplification, which is allowing |rr replay| to be specified with no
arguments.  This replays the "latest-trace" symlink in the saved-trace
dir.

And finally, trace directories are now named based on the basename of
the exe path that's recorded.  So for example, |rr record
/usr/bin/firefox| will create a trace directory "firefox-$nonce".

With these changes, users can do the following

    $ rr record foo
    [Saves recording of "foo" to ~/.rr/foo-0.]
    $ rr replay
    [Replays ~/.rr/latest-trace, which points at ~/.rr/foo-0.]
    $ rr record bar
    [Saves recording of "foo" to ~/.rr/bar-0.]
    $ rr replay
    [Replays ~/.rr/latest-trace, which points at ~/.rr/bar-0.]
    $ rr replay ~/.rr/foo-0
    [Replays ~/.rr/foo-0, which is still saved.]