Refactor syscall_defs.h to contain all known system calls and to encode syscall numbering.
System call numbers are defined via an enum class SyscallsX86 whose members
have the names of the system calls and whose values are the system call numbers.

This change sets us up to add different system call numbers for different
architectures. It also means that when an unsupported system call is encountered
we'll print the name of the system call.