communicate to the debugger stub how big all registers are, not just defined ones
It's conceptually better if, for registers that we don't know the values
of, we don't just say they're 32-bits wide, as the current
print_reg_value does.  Therefore, when we are fetching register values
from the target, we need three values:

- the value itself;
- the size of the register in bytes;
- whether we know the value or not.

To that end, define constants for all core x86 registers GDB might ask
about, and modify the read_register interface to return all the
information we might need.