Move trace_frame member from Task to Session and wrap in current_trace_frame() accessor.
This is actually per-Session state. Currently we store it on the Task
that the event is associated with, but that's not necessary. It's simpler to
have just one per Session. When we restore a checkpointed Session, we can
tell which trace_frame we were replaying when the checkpoint was created.