rename (most) uses of orig_eax to original_syscallno()
This renaming catches the few remaining, interesting instances of
orig_eax.  The ones I didn't rename are either due to be taken care of
by the debugging stub abstraction patches or because the instances are
part of logging messages which print out all registers.  The latter
requires its own, more comprehensive approach to fixing, so I left it
alone.