Ensure that internal breakpoints are removed if and only if we set one.
Also fixes assertion so that we die immediately if a breakpoint's refcounts
go negative.

We were removing an internal breakpoint when we hadn't necessarily set one,
so if there was already a user breakpoint at that location we'd set its
internal-breakpoint refcount to -1. We were also failing to remove the
breakpoint in some cases so gdb would see an int3 instruction there.