Introduce BASIC_TESTS category for tests that use a 'compare_test EXIT-SUCCESS' script.
We're going to have to add a lot more tests as we expand our syscall coverage.
Almost all of these tests will just do "compare_test EXIT-SUCCESS" in their
run script. So it seems like a good idea to have a category of tests in
CMakeLists for which that is the default run script, so these tests don't need
a run script of their own.