Ensure that register-clearing performed by the kernel when entering a signal handler is reproduced during replay.
I'm not sure why the kernel does this, but it definitely does clear FP/SSE
registers when entering a signal handler. The modified sigreturn
test clearly shows this. So rr needs to emulate this during replay. We simply
save the extra-register state in an EV_SIGNAL_HANDLER and restore that state
during replay.

This patch modifies the sigreturn test to dump register values in the signal
handler to check values on entry to the signal handler are consistent between
record and replay. It also modifies the sigreturn test to test x87 FP registers
as well as SSE registers.