fixup SYS_restart_syscall fuzziness
We don't handle SYS_restart_syscall as a normal syscall, so it wasn't
present in syscall_defs.h, and therefore wasn't getting defined in
the *_arch structures in kernel_abi.h.  But it does vary between
architectures and we depend on the value, so we need to define it there.