factor out a prepare_accept function
I think this also fixes a small bug in the accept4 implementation: we're
supposed to determine how much space to allocate for the addr based on
the in-value of addrlen.  The accept implementation was doing that, but
the accept4 implementation was allocating space for a bare |sockaddr|.
Probably didn't matter much in practice, though.