Convert nanosleep to the new infrastructure
This requires adding an option to process_syscall_results to indicate that
the syscall did not write anything and nothing should be copied back out
of the scratch area.

In most cases we could just copy back from the scratch area and mark the
parameters IN_OUT, but this would be less faithful since writes to the
memory locations during the syscall would be lost.