Fix rt_sigqueueinfo and sigsuspend.
rt_sigqueueinfo is broken; it shouldn't record anything, and arg1 is
definitely not a pointer. So we add a test for that, and fix the bug.

Along the way I discovered that our sigsuspend handling has a bug:
if a terminating signal blocked by the process is delivered by
sigsuspend (with a mask parameter that unblocks the signal
temporarily), our signal recording code fails to notice that the signal
was temporarily unblocked, and assumes that the process will be
terminated. So we fix that, and add a test specifically for sigsuspend.