Add infrastructure to emulate ptrace-stops.
During recording we need to detect ptrace-stops in a tracee and emulate them.
Call Task::emulate_ptrace_stop when a stop should be triggered. It will
return true if the task has a ptracer and has been stopped. The ptrace
stop code is stored in the task and the ptracer is notified.

To notify the ptracer, we send it a synthetic SIGCHLD which mimics the
SIGCHLD that would normally be sent by the kernel. We can't pass siginfo
with the necessary values (because the kernel won't let us set
arbitrary siginfo), so we set placeholder values and when the signal
is delivered, we fill in the siginfo with the correct values. The SIGCHLD
forces any in-progress wait syscall to be interrupted; when the syscall
restarts, we will notice if there's a pending ptrace-stop code and if so,
emulate the syscall and return that code. (The same code also handles
the case where a wait syscall is started while there is already a pending
ptrace-stop code.)

The kernel doesn't support multiple pending SIGCHLDs so we avoid issuing
more than one. Instead we track which ptraced Tasks have pending SIGCHLDs
and issue a new synthetic SIGCHLD whenever we're in a good place to
deliver a signal and we have a ptraced Task with pending SIGCHLD.