Refactor ExtraRegisters to eliminate redundant formats.
We really only need one format for x86 --- the XSAVE64 format, of which
XSAVE and FXSAVE(64) are subsets. As noted in the new comments, there
are some small differences in layout depending on the instructions used
to save the data, but that depends only on the kernel, not the ptrace
API we use to get the data (as long as we avoid 32-bit GETFPREGS,
which we do --- it uses the FSAVE format which is totally different). Also
the differences are basically irrelevant in practice.

This simplifies things a lot, which will help us implement ptrace
emulation.