Move runnable_state_changed to run at the start of the scheduling loop
This is quite a complex change, but it will make things easier in the long run.

We introduce Task::has_run_to_a_stop to track whether there is a new stop state
to process or not. This should be cleaned up in future patches.

The goal is to simplify record_step so that we select a task to process,
then if there is a new stop state to process, we process it. During that
processing we decide whether/how to resume execution of the task, and if we
do resume execution, we should (mostly) do that at the end of record_step.
We'll try to reach this goal over a series of upcoming patches.