Close perfcounter fds for tasks in a freshly-cloned ReplaySession.
This conserves long-lived fds. With this change, the test
"Firefox opt build starting up and shutting down, set a breakpoint at
nsCanvasFrame::BuildDisplayList, and then repeatedly do "next; checkpoint""
goes from 161 to 243 checkpoints before failing.

Now, the fds held open for a checkpoint are:
-- one per shared memory segment (in EmuFile)
-- one per thread-group leader (its mem_fd)
The former are difficult to eliminate because we unlink the backing
file immediately after opening it, so we can't reopen it if we
close the fd. The latter are tricky to reopen after closing since
in some cases (Yama security module) we have to do a lot of
ptrace POKEUSER calls to set up the reopen.

We should support enough checkpoints for the forseeable future,
anyway.