Make shared anonmyous mappings work properly.
Shared anonymous mmaps did not have associated EmuFs files, and were
treated as regular shared anonymous maps, so when we created a checkpoint,
the checkpoint shared the same memory as the original, so the original
could corrupt checkpoint state. The solution is to give shared anonymous
mmaps EmuFs files.

This also fixes a different bug, which is that remap_shared_mmap would
open the EmuFs file read-only if the current mapping of that file was
read-only. This is incorrect, because the file might have been opened
read-write and then mapped read-only, in which case a future mprotect
should be able to switch the memory to read-write, but it can't.
remap_shared_mmap can just always open the EmuFs file read-write ---
since we created it, we know it's writeable.