Fix x86-string-coalescing quirk for forward execution
On Intel CPUs, multiple iterations of a REP-prefixed string instruction
may be coalesced so that a watchpoint on an affected memory location may
appear to fire late. (This coalescing is disabled by single-stepping.)
This inconsistency can cause problems for reverse execution, and possibly
for rr users, so ReplayTimeline has fix_watchpoint_coalescing_quirk to
automatically backtrack and reexecute when possible coalescing is
detected. It's trivial to use this code for forward execution too, so
do so. This makes forward execution consistent with reverse execution
and ensures this quirk is never exposed to ReplayTimeline users.