Replace 'syscall' instruction return address in RCX with -1.
The x86-64 'syscall' instruction puts the return address into RCX. When
running under rr, ptrace stops and/or other effects seems to cause
the kernel to usually set RCX to -1 ... but apparently this doesn't
always happen; see issue #1448. To avoid divergence, always set RCX
to -1 in a syscall on x86-64.