Improve `replay_step_to_mark`'s strategies.
In particular, if the breakpoint strategy fails, fall back to single-stepping.
We need this for the blocked_bad_ip testcase, where we can't set a breakpoint
because the program jumped to invalid memory.

To avoid pathological performance issues due to singlestepping through
very many instructions in a long trace event, add another strategy to
`replay_step_to_mark` that uses a ticks constraint to run forwards
until we get close to the mark's tick count. This strategy is actually the
best, once we're in the right trace event; the breakpoint strategy
has a pathological case when the mark IP is hit many times before we
reach the actual mark state. So, use the ticks constraint strategy in
preference to the breakpoint strategy.