During replay, force a task to exit an emulated syscall as soon as we enter it.
This means that tasks are normally never found in the just-entered-a-syscall
state. They can only be in this state for "EXEC" syscalls, i.e. syscalls that
we execute instead of emulating, which are rare. We will eliminate "EXEC"
syscalls soon.

This avoids problems where we would have to speculatively try to finish
an emulated syscall when there wasn't one pending, which can trigger
unexpected code execution, and even trigger the kernel restarting
a system call that shouldn't be restarted (which is tested by
`clone_interruption`).