Assert in Task::write_mem that the struct we're writing does not have holes.
Holes (i.e. gaps in the memory layout where the compiler has inserted
padding between fields) in a struct written by rr to tracee memory are
extremely dangerous, because those will likely be uninitialized in the
rr process, so uninitialized rr memory will probably be written to the
tracee. The contents of this memory can vary across rr runs, so this can
cause divergence later in the tracee if it uses that data.

The assertions inserted here do not guarantee that T has no holes, but
they can catch some bugs.