src/AddressSpace: Handle PROT_GROWSDOWN in mprotect()
The flag PROT_GROWSDOWN passed to mprotect() means that if the mapping
being changed is a grows-down one (ie stack) then the change should be
extended down to the start of the segment. Implement this behaviour,
and make sure we don't pass the GROWSDOWN/UP flags into the mappings
we create (as this confuses code which reads them later).