Emulate MADV_DONTFORK during replay
Our checkpointing code needs to ignore MADV_DONTFORK, but our real forks need
to honor it. It's a little simpler and perhaps more performant to make
MADV_DONTFORK emulated during replay, i.e., we don't tell the kernel about
MADV_DONTFORK and instead we track the DONTFORK regions and manually unmap
them after a fork.

While we're at it, make all madvise calls fully emulated. In general
we prefer emulating syscalls during replay since it simplifies what
the kernel is doing and avoids having to deal with states where we
need to checkpoint a process that's currently in the kernel. In this case
MADV_DONTNEED may not work properly/deterministically when executed
since it's documented that DONTNEED on file-backed pages may throw out
dirty pages and re-read data from the underlying file.