Support SECCOMP_RET_TRAP in seccomp-bpf filters.
The Firefox (Chromium-derived) sandbox needs this. To make this work we follow
the same strategy as for SECCOMP_RET_ERRNO: our filter rewriter translates
SECCOMP_RET_TRAP to SECCOMP_RET_TRACE with a special data value
(EMULATE_RET_TRAP), and RecordSession then emulates the SECCOMP_RET_TRAP
behavior. This means that for AutoRemoteSyscalls performed on behalf of rr,
the seccomp filter runs and generates a PTRACE_EVENT_SECCOMP which is ignored,
so our syscalls work no matter what the filter is.

RecordSession's emulation generates a Syscall event for the triggering
syscall, and then stashes a SIGSYS with the correct siginfo for later
delivery to the tracee. To deliver the correct siginfo without it
being sanitized by the kernel, we have to copy it to the signal handler's
siginfo_t* after we've entered the signal handler.