Make SYS_rrcall_notify_syscall_hook_exit set its return value and original_syscallno to specific values passed in by the caller.
This ensures that any signal which was intended to be delivered immediately
after a syscall that was invoked while signals were disabled, and which
is delivered immediately after SYS_rrcall_notify_syscall_hook_exit instead,
sees register state that looks like we just completed the original system
call. Furthermore any changes to the syscall result performed by the
signal handler will be returned to the caller of the patched syscall.