Create "privileged" syscall entry points so that syscallbuf's helper syscalls aren't intercepted by tracee seccomp filters.
When using syscall buffering in a tracee sandboxed with seccomp filters,
our helper syscalls such as the syscalls that arm/disarm desched events
are likely to be blocked by the seccomp filters. We work around this by
creating syscall instructions in the rr_page that are treated as
"privileged": every tracee seccomp filter is prefixed with code that
always allows syscalls from those instructions. Our syscallbuf code uses
these privileged entry points for all syscalls except for the syscalls
it performs on behalf of the tracee program.