Handle sigprocmask being called during a signal handler running while the syscallbuf is locked.
While a signal handler is running that has interrupted a syscallbuf syscall,
we keep the syscallbuf locked to prevent reentry that would corrupt syscallbuf
state. If the handler calls sigprocmask, it's totally legitimate to see that
the syscallbuf is locked there regardless of which signals are being blocked or
unblocked, so the assertion we had in update_sigmask was bogus.

update_sigmask was trying to set the syscallbuf locked state if the necessary
signal was blocked. A better approach is to prevent the tracee from blocking
syscallbuf's desched signal (and rr's timeslice signal!) in the first place.
This patch does that. Since update_sigmask no longer needs to know anything
about syscallbuf locking state, we can remove the problematic assertion.