Rewrite AddressSpace verification logic
There are a few changes made together here.

Using the new-ish /proc/<pid>/maps iterator, we can restructure the
verification logic to be much clearer: coalesce all continguous kernel
mappings, then coalesce all contiguous cached mappings, check that the
two results match, then move on to the next chunk of the address space.

The other big change is to store in Mapping.map the real device and
inode numbers instead of the recorded ones, and update EmuFs to look
in recorded_map for the device and inode numbers during recording. This
should massively reduce confusion.