Emulate execve() syscalls during replay.
Instead of exec'ing the original binary, exec a stub program, unmap all
the created mappings (except for the rr-page and [vsyscall]), and map
in all the mappings we recorded after the original execve call.

We still have to do an exec because sometimes we have to switch a task
between 32 and 64-bit architectures, and an exec is the only way to do
that.

Handling execve this way lets us use our regular file-mapping machinery,
so if the exec'ed binary changes or disappears we can handle that properly,
and we have the option of copying the exec'ed binary to the trace.

This will also let us reenable ASLR for rr tracees, improving the entropy
of tests running under rr.

The behavior of 'brk' and 'mremap' depends on hidden kernel state that
depends on which binary you exec'ed, so now that we're emulating exec,
we need to emulate 'brk' and 'mremap' too.

This all gets us closer to the goal of emulating all syscalls.